Version 1.0.0 Working Custom Appender. Changes in the configurations.
Implemented for each Root log to call the custom appender which configured in the logback.xml file.
When logger.info or warn or debug or error is being used through logback (configured in every project with the custom appender inside) then an object is being generated for a post request in the other project, this way the logs are being stored in the db.